[lsp][refactor] Allow non-exact selections for expression and type refactors
Summary:
Almost two years ago, I built the extraction refactor code actions. However, I left over a todo that make the feature less discoverable: you have to do exact selection of expressions and types in order to make the refactor options pop-up.

This is certainly not ideal, and it's a result of lack of information from AST. Consider the following example:

```
const x = 'foo';
//      ^^^^^^^
```

The selection includes some extra stuff, so we definitely should not trigger an extract expression refactor. However, at the AST level, this is indistinguishable from the following case, which should be allowed:

```
const x = 'foo';
//       ^^^^^^
```

Therefore, at the time when this is implemented, I decide to not make this block the feature, and require exact selection instead. However, from user feedbacks, it's clear that exact selection makes the feature less discoverable, so we need to fix it.

We concluded above that this is impossible to do on AST, so we need to remove some layers of abstraction to get the information we need. It turns out that, the only other info we need is the token list. With the token list, we can easily tell whether the selection range also includes other tokens that should disqualify it as an available refactor.

We can obtain the token list by reparse the source with `token_sink`. A linear scan of the token list is indeed expensive, but we only need to do it at most 2 times (one for expressions and one for types), which should be OK and still much less expensive than type checking.

Changelog: [feature] To use the extract to expression/type alias refactors, you no longer need to do exact selections to trigger it. Any selections that include extra whitespaces will be OK too.

Reviewed By: darichey

Differential Revision:
D45251081

------------------------------------------------------------------------
(from 4216df126cf8a386a137d244895282b8b3589f52)

fbshipit-source-id: a1a245e06c16cf3cf0429eec89a17acdab0de106