add LatentPred variant to Type.destructor
Summary:
My goal is to remove the distinction of `defer_use_t` into `LatentPredT` and `TypeDestructorT`
and only keep the latter. `LatentPredT` is the internal type for the `$Refine` utility type. It is not
officially announced and so is practically unused. This diff makes it behave like all other EvalTs.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D45342065

fbshipit-source-id: c5f6f264d491f2f63cf4149a4d3a57d236343ab6