[errors] Show the top scoring hidden branch (if one exists) if we were otherwise only going to show a single branch
Summary:
The intention of the our error branch scoring and hiding logic is to avoid showing very large and unwieldy errors.
However, we often hide the wrong branch. Ideally, we would perfect this branch scoring logic, but in the meantime a simple change would be to show the top scoring hidden branch (if one exists) if we were otherwise only going to show a single branch.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D45326215

fbshipit-source-id: a2e39b32f94d93e7e390eec73407a1169692456a