in type-at-pos command serialize Ty.t at the server
Summary:
This diff changes the type of `infer_type_response_ok` to not include`Ty.t` types in
```
tys: Ty.type_at_pos_result option;
```
The problem with this type is that it is sometime awkward to serialize `Ty.t` on the client since these
types might include `ALoc.t` that are not concretized.

Instead, the new type is a
```
type infer_type_response_payload =
  | Infer_type_string of string option
  | Infer_type_JSON of Hh_json.json
```
which does not include `Ty.t`.

Conversions of `ALoc.t` to `Loc.t` and then `string` happen on the server-side which has access to
aloc-tables. See change in ty_debug.ml.

The rest of the changes are mostly mechanical. I introduced a `Infer_type_options.t` type to avoid
repeating some parameters.

Changelog: [intenal]

Reviewed By: SamChou19815

Differential Revision: D45366373

fbshipit-source-id: 38fe410e2646822c63921d4879a9c1c22a6aaf39