Add separate definition for root with specific reason case
Summary:
For some root ops, we want to use one location to check for containment, but
point to a different reason.

For example, for function calls, we want to use the location of the entire call
to check for containment, but the function name as the root location.

Before this diff, passed an optional specific reason to all callers of `root`.
This diff factors out a separate `root_with_specific_reason` definition and
removes the optional param.

Reviewed By: SamChou19815

Differential Revision: D45402111

fbshipit-source-id: 7ab662f4abbe34ee9bae39a4b4de10550f11cb4b