Change error concretization/filtering in optional props codemod
Summary:
I'm trying to optimize error collation, which affects the concretize -> make
printable -> filter suppressed codepaths.

Concretizing entire errors can be expensive, so I'm working to fuse these
operations, so we can filter suppressed errors before we do lots of work to
concretize them and make them printable.

This codemod needs to look at the non-printable error and get concrete
locations, but only if the error is not suppressed. This diff makes a few
changes, which unblock later work:

1. Use the full loc_of_aloc function which will not fail with Not_found. This
avoids the need to handle exceptions.

2. Get the locations needed by the codemod from the error with abstract
locations, using loc_of_aloc as needed, instead of relying on a fully
concretized error.

Reviewed By: SamChou19815

Differential Revision: D45417032

fbshipit-source-id: a9c85c9052eb5a8593bc6db28acb34a72deac7eb