Simplify property compatibility frame unwrap
Summary:
The logic for this is pretty complicated. I'm not sure it really makes sense.
But regardless, this change doesn't affect the logic itself. Instead of passing
around a reason, we pass around the reason's loc.

Before this change, the reason was passed to `unwrap_frame`, which only used the
loc anyway. This diff adds `unwrap_frame_with_loc` so we can avoid repositioning
reasons and just accumulate the loc itself.

Reviewed By: SamChou19815

Differential Revision: D45433989

fbshipit-source-id: f283bd9b951a474e2633ada23d5af67e671c1d9d