Use combined API for concretizing and making single errors printable
Summary:
Similar to D45417046, but for make_error_printable instead of the multiple-error
case.

The annotation on make_error_printable is required because this function
(temporarily) requires polymorphic recursion.

We still fully concretize errors, but this is now an internal detail, which we
can change. Later in this stack, we will successfully avoid visiting most of the
error structure in most cases.

Reviewed By: SamChou19815

Differential Revision: D45418343

fbshipit-source-id: dc46aa352834a601e3648f58622d81a3fb86b944