Replace recursive mapping of speculation error branches
Summary:
When we print a speculative error, which has nested errors in its branches, we
want to use the code of the top-level error.

Before this diff, we accomplished this by mapping over the speculative error,
replacing all error codes along the way.

Instead, this diff threads the top-level code through a later traversal, and
simply ignores error codes from speculation branches.

This accomplishes the same result without mapping over the speculation error,
re-allocating the spine of the error structure.

Reviewed By: SamChou19815

Differential Revision: D45504887

fbshipit-source-id: 3025d0f885cdf11c21db774753ce315020d71b09