Move expensive speculation branch deduplication after filtering
Summary:
We deduplicate branches in speculation errors (see D14446492). Currently this
happens in Errors.mk_speculation_error, which is called from
make_error_printable.

After we make errors printable, we generally filter suppressed errors. In roots
with lots of suppressed errors, the cost of deduplication can add up.

This diff moves the deduplication logic later in the error pipeline, after we
filter suppressed errors. As a result, we don't pay the cost of deduplication
for suppressed errors.

Reviewed By: SamChou19815

Differential Revision: D45509093

fbshipit-source-id: c95a68a812f0e0c17a73933c1c0f53b503675051