accept any expression as predicate function callee
Summary:
Before, the predicate functions callee `e` in `if (e(x)) {}` needed to be either an identifier or a
member expression. Event hough this accounts for the vast majority of the cases in practice, there's
no fundamental reason we need to be limited to these two forms.

This diff relaxes this condition. It changes the `func` portion of `LatentR` to any expression. The
dependencies of this expression are recorder in name_def_ordering.ml as usual. The resulting type
of the callee expression is stored in a lazy value in a new map in Loc_env.t, which is accessed in
env.ml when the `LatentP` predicate is created. The reason for this indirection is that env.ml, where
this conversion happens, does not have access to statement.ml functions.

Notice the use of `with_allowed_method_unbinding` to bypass method-unbinding errors on the
top-level of `e`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D45509946

fbshipit-source-id: cbeec45cb3f25043724defb95fafdb82c3dcf931