add "allow-method-unbinding" bypass
Summary:
Adds a `with_allowed_method_unbinding cx loc f` function in Context that allows `f` to be run
with `loc` being an immune location to method-unbinding errors.

This will be useful in a follow-up diff where in code of the form `if (e.m()) {}` we would like to allow
checking `e` with a general purpose `Statement.expression` but want to avoid erroring on the
top-level expression if it happens to be a member access. In this case the location that will be passed
to `with_allowed_method_unbinding` will be that of `e.m`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D45509945

fbshipit-source-id: 335c7a0a2a86b8f8de5e1cbc9e6c592d6bb5f580