fix autocomplete token inside comments
Summary:
previously, the `ac_loc` (loc of the token being completed) when completing inside a comment was the entire comment, even though the token was a specific word inside the comment. the token's actual loc was a field of `Ac_comment`.

we measure autocomplete "sessions" by how many consecutive requests are for the same token start loc. when the token loc == comment loc, anything you write in the entire comment is one session, even though each word is a separate token.

instead, this swaps the two locs: `ac_loc` now matches the token being completed, and the entire comment's loc is nested in `Ac_comment`.

everything works exactly as before, except:
- this fixes the computation of `session_requests`
- it includes the sigil in our FB-only telemetry
- $FlowFixMe completions now "insert" instead of "replace" by default, but you now have the option. if there's existing text in a comment above an error, and you manually invoke autocomplete, the `$FlowFixMe[foo]` suggestion will be inserted in front of the existing text instead of losing what you wrote; you can also choose to "replace" the existing text.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D45512141

fbshipit-source-id: d0f12d1d26ee78a1fd4557250f0012c620b02e5b