fix autocomplete session length
Summary:
suppose you type `foo = bar`.

space is an autocomplete trigger character, because we use it in JSX; we have to ignore it whenever it triggers. so when you have `foo = |`, we return an `Empty` ac_type and ignore the completion request because it should only trigger once you type the first char of the word `bar`.

once you type `b`, the ac_type is `Acid` (AutoComplete IDentifier).

in both cases, the token starts at the same place.

the bug being fixed here is that the autocomplete session counts consecutive completion requests with the same token start loc, which results in `2` in this example -- one for Empty, one for Acid -- but the Empty one shouldn't count.

the fix is to also require the ac_type to match, so the transition from Empty to Acid starts a new session.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D45514561

fbshipit-source-id: f25633350e56af3ed87db088bb87131776a0bdcf