explicitly ignore autocomplete on type bindings
Summary:
just as we can't suggest anything for `const |` because we can't make up a variable name for you, we can't suggest anything for `type |` because we can't make up a type name for you.

we handled this already, but only by making sure the AST search for the autocomplete cursor doesn't find anything.

instead, we should explicitly handle this case so we can (a) stop searching for the cursor, (b) properly classify the ac_type and avoid a "FAILURE" result

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D45515279

fbshipit-source-id: 5209db62f235914ac0c4787fcd57b3ad0e625179