fix in type_sig-parsing of predicate functions
Summary:
Fixes the case of exported signature for this code
```
const y = "abc";
declare function mono(x: mixed, y: mixed): boolean %checks(typeof x === "number");
export function wrapMono(x: mixed): boolean %checks {
  return mono(x, y);
}
```
The presence of `y` in the call to `mono` (which is not a parameter to `wrapMono`, would throw off
the parser and made it ignore every refining position.

This diff restores this behavior to instead just ignore the non-parameter id-argument to `mono(..)`.
This now matches what happens when `wrapMono` is called in the same file.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D45558707

fbshipit-source-id: bed7f756c0feaee38dd6b1bed3c66dd8b695f888