Allow opaque types with string supertypes to be extracted with as keys using Object.keys
Summary:
Changelog: [feature] Allow opaque types with `string` supertypes to be used as keys in a dictionary, and have that opaque type be preserved when using `Object.keys`

This is done by preserving the original type when using `ToStringT` with opaque types which have supertypes. I'm open to ideas for a better approach!

Reviewed By: panagosg7

Differential Revision: D45332712

fbshipit-source-id: 949ef84015d8d35f0583b478fbfa8f88d359039d