Autocomplete jsx intrinsics
Summary:
When autocompleting in a jsx tag, include members of `$JSXIntrinsics`.

Note that this diff does not improve the types of those intrinsics, so we only get the names of the elements for now. When we improve those types, we will get better autocomplete for their props automatically.

Changelog: [ide] Added built-in element names to JSX autocomplete

Reviewed By: mroch

Differential Revision: D45003815

fbshipit-source-id: 0bf392b4540dffed7f3a835b2a93aef9e077550e