[mapped-types] Add homomorphic flag to mapped type destructor
Summary:
This diff adds a homomorphic flag to mapped types and sets it to true when there is an inline `keyof` in the mapped type definition. It's impossible for us to be fully consistent with TS here because TS will eagerly merge unions of object types to the intersection of their keys. This is not something we do, and that would be a massive breaking change.

I chose to go with the inline `keyof` rule for determining if a mapped type is homomorphic because:
1. It appears like that's how TS works
2. that design is very simple to explain-- if you want distributivity in the result of the mapped type use an inline keyof. This also lets us avoid really weird scenarios, like:
```
{[key in (keyof O) | 'foo' | 'bar' | 'baz']: number}
```

What should the behavior here be? Consider keyof O separately and then flatten the rest of the union? TS just flattens the union here and does not attempt to be distributive.

[See this TS playground link for more examples with discussion](https://www.typescriptlang.org/play?#code/C4TwDgpgBA8gjFAvFA3gMwPYYFxQHYCuAtgEYQBOANFORAIYAmGeANiFCXeboaRQL4BuAFChIsAExJUtRszYcuPYmXJDhGgPSaoACQxEDGcmAAWASwDGASVbm80ZCllNW7Tt3wqBAOigARcwBnYHJzEgJgCAYoDAA3CihgU2gCPHNmKDo8GKIKAHNopIwoACsCEKSUqHso8iCIS2AMvFi0KABrCBAgn2FtKAAxFgwAdyhRjAIWGLByDAYCS2h0LGU+KigAag911X4oAB9UHaUvDf4-ADkMZuWqumAoQwZzNHMKIKzaKDmIBvICQYonA0H0hkMJgsNjsDmkKAA2l12PZOt0MO0ABTwI6SACUAF09gJBFANGJoABpbpfZDIjFQbEIY4wCR4kQDG7AcFGKFWAAqYUsHXYTl2532fhxLKkwSg9EK5AUQXMRDALDeHxiwBKKHFvH21CCJXpWOl+OkAHIPJbrrcrNBko8aPRXAo5Qx5mBIAw+hSoFyeZCzAKhSL4UjujVWtSekSJSStDoADIQYCWr6vEJhCJRWIJchVVLpTJEbIEOgsNjYEHiWNBBB09HteAif31qRNkAM1lt0EB25B4whyz+f5gcxRACiBZAyXs+QAqiq8PlYwynJGetGoPXmbuaRJ4waSVABk5MDgE5t9d41NQ8CUXPIQNQ0i0oJYMFW6GAGsD+h0QZjCgfIsFyegggIWhqBYNMM0-Zh3nIIgHieLNQnCSIF2jDU4RCcwq0Qn8-3+IsoHfZg-HMHwIGoIhpmadVoFNWl8FuPQjGDaFa2gQMuOHaFAmzLCogYWw8McVAty+VFMVNWA4DxXF5ObfFjzvIQzx0C81mvahbwuPx+WqSjWlGOgvi-Ej-yyHJ2KgDCc2wuJJ3YUxf0gPBeigABZBZNU+dinj+AEgR8IA)

This diff does *not* implement support for non-homomorphic mapped types, but the rest of the stack does.

Reviewed By: panagosg7

Differential Revision: D45447495

fbshipit-source-id: 856190ce057a4f29f5f75c1a1d726e9fe90a495a