Tweak imported module type queries
Summary:
Before this diff, we used import/import_ns, which both queried for the imported
module type. Due to the structure of the code, for an import statement with
multiple specifiers, e.g. `import { foo, bar } from 'baz'`, we would query for
the module type for `'baz'` twice: once for `foo` and once for `bar`. We would
do similar things for export specifiers with a source.

This diff changes the Import_export API so that `import` returns the module type
and `import_ns` uses the module type to get the namespace type. It also queries
for the module type once per import/export statement and passes the module type
down, instead of querying from the leaves.

This is a cleanup, but also prepares for adding module types to the typed AST
for export statements.

Reviewed By: panagosg7

Differential Revision: D45590986

fbshipit-source-id: 0a4fc1c8b600f97f3dd632dd98055c7eeac2c082