Add literal AST node for haste module refs
Summary:
Flow supports "haste module refs" which are string literals with a specific
prefix. These are treated as references to modules instead of strings when code
is bundled.

Currently, we detect module references during AST analyses, by adding special
handling code to string literal AST nodes.

This diff moves the detection of this special case into the parser. Instead of
creating a string literal AST node, we create a new "ModuleRef" literal value.

This makes it easier for AST analyses to handle the module ref cases, but my
primary motivation here is the added `module_out` field, which will contain the
referenced module's ModuleT type. This will allow consumers of the typed AST to
conveniently get the module type.

Reviewed By: mroch

Differential Revision: D45612755

fbshipit-source-id: 874123c3ad66ab999e46d3e4990986ab4bb65e43