[errors] Use RMixed as the reason description of tparam default bound
Summary: Changelog: [fix] For error messages produced as a result of incompatibility between a generic type with the default `mixed` as its bound and another type, we will display `mixed` instead of the name of the generic type. We hope this improves you understand the cause of the error better.

Reviewed By: panagosg7

Differential Revision: D45592595

fbshipit-source-id: a0f1d000ab009a0698fa1fecf7ca22d4802ead00