[mapped-types] Add more concretization for mapped type arguments
Summary: The Hint concretization machinery is not interested in splitting up $Keys. Mapped Types certainly want to break up $Keys instead of using the $Keys as an indexer. This diff adds a new concretization target and adds the relevant case in flow_js.ml to make sure we preprocess $Keys.

Reviewed By: SamChou19815

Differential Revision: D45447494

fbshipit-source-id: d0b41d8877b559d6782849d3e58113f9be6b7fc5