[mapped-types] only allow string | number | symbol keys
Summary:
[This restriction matches TS](https://www.typescriptlang.org/play?#code/C4TwDgpgBAsghmSATKBeKBvA2gawiKASwDsoAjAewoBsI5iBdALimIFcBbMiAJwF8A3EA).

Making sure the error message looks good is tricky here. Since the upper bound is a union, we get a speculation failed error instead of an incompatible-use error. This is the first time we special-case a speculation failed error. The special casing is necessary because there is no syntactic `string | number | symbol` type to point to in the incompatibility message.

Reviewed By: SamChou19815

Differential Revision: D45548638

fbshipit-source-id: f5ad4bc3dff426565c8b91eeaf2f91118d7dc9f8