[mapped-types] Non-homomorphic mapped type checking
Summary:
This diff adds support for non-homomorphic mapped types.

There is one detail worth calling out: The normal EvalT machinery has special handling for unions, where it will apply the EvalT to each member of the union instead of considering the union as a whole for input. This is incompatible with how mapped types work. Mapped Types need to process and flatten the entire union, and only then can it be computed. To flatten the union, I use the same PreprocessKit machinery used during hint evaluation. This is not without drawbacks: if any of the tvars we encounter are truly unresolved then we may miss keys in our output type. However, this problem is not unique to any specific approach: if we didn't resolve the type immediately and instead kept re-triggering the evaluation on each new bound then we would infer unions of objects instead of collapsing all the keys into a single object. There is no way to achieve the desired behavior in 100% of cases with Flow today, but with LTI the set of bad cases is extremely small. I was not able to construct an example that exhibits bad behavior, including in implicit instantiation cases.

This diff enables Pick/Omit, which are included in the added tests.

Reviewed By: panagosg7

Differential Revision: D45447493

fbshipit-source-id: 5637a19633562be4255da9ee0b086bf04c155492