add targs and arg_ts to CallLatentPredT
Summary:
Motivation: Type-guards will rely on the CallLatentPredT mechanism. Unlike `predicate`-based
refinements, type-guards also rely on the typing of other arguments of the function. Consider for
example
```
declare function pred<A, B>(x: mixed, y: B): x is [A, B];
declare var x: mixed;
if (pred<string, _>(x, 42)) {
  x;
}
```
The concrete type of the guard is determined by the (type) arguments to the call to `pred`:
* `A` will be `string`, which is explicitly passed in
* `B` will be `number`, which is inferred from the parameter `42`

This will allow us to infer `x` as `[string, number]` in the then-branch.

To enable this `CallLatentPredT` needs to know about the type arguments and argument types
of the refining call. This diff adds the necessary plumbing to do this.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D45558879

fbshipit-source-id: c9ad4e6a65fdceadccc58a2f19744c61ee982197