early stop name-ordering traversal
Summary:
The next diff will add edges to the name-def ordering graph that connect the refined version
of a predicate function call to the arguments of the call.

In code patterns like the following
```
if (Pattern_1.test(x)) {
  x;
} else if (Pattern_2.test(x)) {
  x;
}
...
} else if (Pattern_n.test(x)) {
  x;
}
```
the refinement reaching the last occurrence of `x` (lets call it `x`_n) will include edges to the call of
`Pattern_n.text(x)`, but also the negation of the previous branches. The previous occurrences of
`x` (`x`_n-1) would itself depend on the a similar set of locations, all of which though are
subsumed by those of `x`_n.

It therefore makes sense to end this traversal (in `use_visitor`) as soon as we discover a node
that has been visited.

This diff does this for expression nodes, that are the perhaps the most common in this chain.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D45598628

fbshipit-source-id: a91a9477d69a0d58399432ad8f874b8f0d7ebb71