Use require type instead of module type for module ref literals
Summary:
I added this AST node in D45612755. In that diff, I made the typed AST reference
the ModuleT type. The plan is to use this type for get-def, but it turns out
that the t_out from CJSRequireT is actually more useful, since we want to jump
directly to the CommonJS export, if there is one.

Reviewed By: SamChou19815

Differential Revision: D45637607

fbshipit-source-id: d947419167804aea7afc74dd425c1f758bb54602