Don't check `@noflow` files in "all" mode
Summary:
Flow can be configured to check all files, even those without the `flow`
pragma. We use this for tests, mostly.

Before this diff, Flow would even check files with a `noflow` pragma. This
makes it difficult to write tests expecting an un-checked file, which is
sometimes necessary to test such cases.

We could simply not use "all" for such tests, and I'd be happy to abandon this
diff if that's the right path forward. This seems like a reasonable behavior,
though, and let me keep "all" for more tests.

Reviewed By: mroch

Differential Revision: D45628959

fbshipit-source-id: 46856eba7b857037ebb8808e1d37ef2baa4705e2