Use typed AST for get-def on require() and module ref literals
Summary:
Instead of the custom logic for Get_def_request.Require, we can now just use the
type from the typed AST.

There is one behavior difference: for requires which fail to resolve, we used to
return an error response, and now we return a successful response, but point to
the require itself.

I think it might be better to return no result, but the existing APIs make it
difficult to achieve that result. Between a spurious error and a spurious
result, I think the result is better.

Reviewed By: mroch

Differential Revision: D45637608

fbshipit-source-id: a5dc6441dfe656e1354b68486b19254152f2d053