Reposition any-typed modules through import uses
Summary:
For import/export-from/require, the source can have a AnyT type if the file is
unchecked or coresponds to an unknown file. When we encounter these cases, the
import use's type is also `any`.

Before this diff, the import use's `any` was a fresh type. This diff instead
repositions the input `any`. The important difference is that a repositioned
type maintains the input's location as the "definition" location, or def loc.

The def loc is used by IDE commands like get-def. Currently get-def has special
cases for require() calls, to ensure we jump to the right place. Making this
change unblocks some work later in this stack, which eliminates these special
cases.

Reviewed By: mroch

Differential Revision: D45629684

fbshipit-source-id: 52ea78e385de20c9d075e294bfb69d77ba2eeb17