[mapped-types] Use ObjKitT provided reason for result of a mapped type
Summary:
Instead of using the reason from the slice, the mapped type result should use the reason passed into ObjKitT, which is the RMappedType reason created in type_annotation.ml.

This diff is not just improving error messages, but it also affects safety when the objects passed into the mapped type have a literal reason. We want to be sure the output is treated like a non-literal. I discovered this issue when the "Semi-homomorphic mapped type checking" diff failed an existing test case because the reason used a literal reason.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D45694420

fbshipit-source-id: e6024fb5a70eed7fc075afa03fecf406b82567ab