reset TypeAppExpansion stack in run_and_rolled_back_cache
Summary:
Before, we allowed implicit instantiation to reuse the TypeAppExpansion stack. This means that it
was possible to encounter a `T ~> TypeAppT (OpenT _)` constraint, during the instantiation that
would hit the cache and return early. Effectively, for implicit instantiation this means that the OpenTs
within the TypeAppT could remain under-constrained. See added test for example of this.

This diff addresses this by reseting the TypeAppExpansion stack in `run_and_rolled_back_cache`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D45683693

fbshipit-source-id: 6ca33294ae29003f8d8908554745d3d111ccee6c