use_op in CallLatentPredT
Summary:
In the next diff we will be treating `_ ~> CallLatentPredT _` constraints as an actual call
at least as far as implicit instantiation is concerned. This means that it is possible for it
to introduce [underconstrained-implicit-instantiation] errors. As such it is important to
have these errors match the ones that the respective `_ ~> CallT _` constraint will introduce.

An example that would cause this error with both constraints is
```
declare function fromMaybe<T>(x: mixed, ?{ f: T }): T;
if (fromMaybe(x)) { // raises [underconstrained-implicit-instantiation]
  (x: mixed);
}
```
Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D45686370

fbshipit-source-id: 26bf2b4ae4a9f6f001f605ca532febcd68d6ed65