Extract Test_utils module for creating options record
Summary:
The options record is a required parameter for a ton of functions which are
interesting to test. It's useful for tests to construct this value, but
cumbersome to do so.

This diff extracts some existing code into a new module which can be used from
other tests. I made the API more flexible, so tests could more easily create
options with specific values.

Reviewed By: SamChou19815

Differential Revision: D45738191

fbshipit-source-id: 6086f93fdbdbfaf90d4b2037679fd1910b90bb32