[mapped-types] Semi-homomorphic mapped type checking
Summary:
This diff implements type checking for semi-homomorphic mapped types. Given a definition like:
```
type Pick<O: {...}, Keys: $Keys<O>> = {[key in Keys]: O[key]}
```

The approach we take to preserve property modifiers/distribute over unions is to emit an O ~> ObjKitT (MappedType Keys) constraint, which will convert O into a resolved slice. Given a mapped resolved slice, we pick the subset of keys to include in the result based on the resolved type for `Keys`. These object types can either be exact if no indexer is included in Keys or indexed if one is.

Reviewed By: panagosg7

Differential Revision: D45656817

fbshipit-source-id: 98fd19ccd9863e0a8e38b6f6ec5792881814195e