[mapped-types] Semi-homomorphic mapped types normalization
Summary:
This diff adds in normalization for semi-homomorphic mapped types.

I augment the homomorphic field to include a SemiHomomorphic variant to accomplish this, and most of the changes are just plumbing from there. We make use of the variant in the ty_serializer to make sure that we report the correct source type in the typed ast.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D45673279

fbshipit-source-id: dd943401e0653d55c11e8b8f75f41dd6c3f496c1