[mapped-types] Change homomorphic field to inline_keyof in type_sig for mapped types
Summary:
Instead of tracking if a mapped type is homomorphic, which depends on semantic info for semi-homomorphic types, we now only track if there is an inline keyof in the definition. In the diff stacked on top of this, we handle semi-homomorphic mapped types in signatures by inspecting the source_type we have access to in type_sig_merge.ml.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D45673278

fbshipit-source-id: 557aebadb80aa298c44d8bd21132fb0ac75f7eb1