Simplify Parsing_service API, remove Parsing_options
Summary:
Instead of accepting Parsing_options, this diff changes the Parsing_service code
to accept the Options.t record. This is the value that we pass to most other
phases of the type checker: init, merge, check, etc.

Parsing_options is mostly derived from Options, with the addition of some logic
to force "types" mode and force "use strict" mode. For these cases, this diff
uses optional arguments, which I've also renamed for clarity.

I've wanted to improve the Parsing_service API for a long time, but my immediate
motivation here is to make it easier to call into this code from Init, where we
parse lib defs.

Reviewed By: SamChou19815

Differential Revision: D45748232

fbshipit-source-id: 5f046a6e4f8126da036c623ce3fd1c77e0e54d53