[lsp] Convert location to lsp location directly for find references results
Summary:
For global find refs, the assumption that the document uri is the same as the source document uri no longer holds. We will need to actually derive the info from locs. I also renamed the functions to no longer imply that the find all refs are limited to local ones.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D45755001

fbshipit-source-id: 59f6d348c658a71d0fccbb02f572fd0824839633