[mapped-types] Distributive mapped types in signatures
Summary:
We distribute over mapped types if the source is a GenericT. This diff adds the same behavior to exported objects. Prior to this diff, the added tests marked OK would fail.

To achieve this, we simply check the source type after we figure it out in type_sig_merge. There is no reason for us to add any new information to the type_sig representation.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D45800182

fbshipit-source-id: 246402b473df479d9d1a73649b04e387dabe1ccd