[mapped-types] Distributive mapped types
Summary:
Like conditional types, Mapped Types distribute over unions. In order to achieve this we use the same approach that's taken with conditional types and add a distributive_tparam_name to delay the substitution. This diff uses all the same helpers as the conditional types to achieve the goal.

I could not separate normalizer changers out into a different diff because without the changes the existing tests fail. The lines added to the normalizer don't feel quite right, but I couldn't figure out another way to achieve my goal.

The next diff adds support in signatures for this extra field.

Reviewed By: SamChou19815

Differential Revision: D45799306

fbshipit-source-id: 9239035f207a2dac442e47904685eafbd5910d39