[mapped-types] Do not use TypeAppT ~> TypeAppT special case when comparing EvalTs
Summary:
When two type applications flow into each other flow attempts a performance optimization to avoid expanding out the aliases. Instead of computing the full types, when two equal aliases run into each other we just compare the arguments by their polarities.

This does not make sense for evaluated types, where those arguments do not reflect the actual meaning of the type in the alias. This can cause things as simple as a user-defined ReadOnly to be incompatible with itself when writing out the evaluated types manually would not be. See the added test case for an example.

This diff prevents mapped types from participating in this perf optimization when the top-level type in the TypeAppT is a mapped type, which is a good 90% solution. Ideally, we should get rid of this case entirely, but there is a lot of code that depends on it.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D45816760

fbshipit-source-id: 0d9138706c87a14a4a0b4eb638a62a661e8f1b13