[find-all-refs] First cut of find all references for props
Summary:
This diff implements the initial prototype of find all references for properties, gated behind the flag of  `experimental.global_find_ref_props`.

The task can be divided into several tasks:

1. Jump to the definitions of the references
2. Find what files that can potentially contain the references
3. Check those files and find references on the typed AST
4. Collect all results and show them to the users

Task 3 is most straightforward. Once we have the file list, we will run a check to get the typed AST. We thread through the find refs request so that the check service will optionally do some find all ref computation. The results are merged together in a similar way as error set.

Task 2 is achieved by building upon the dirtifying signature work in the previous diff. During merging phase, we use a flag `for_find_all_refs` to decide whether we care about whether type sig entries are dirtified. If the flag is set to true, then we will dirtifiy the sig hash by inverting the bits. If not, then we do nothing special. Note that these dirtified sig hashes are never written back to the shared memory. Why? Because all we care about is that we will force the rechecker to think a file must be rechecked. After the diffing with the old sig hash is done, the new sig hash serves no other purposes. Storing the sig hash into the shared memory will cause the second of two back-to-back find all refs to return empty results, because it sees the exact same dirtified hash as the previous one.

This is an initial prototype, so there are still a lot of things that haven't been completely ironed out. e.g. canceling the find-all-refs, which I plan to gradually address in later diffs.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D45757052

fbshipit-source-id: 713432518c46ac3283f2750481efb24b5ca053e7