[type-sig] Dirtify entries for find-all-refs
Summary:
This diff adds the infrastructure to perturb the type signature by marking definitions associated with some locations dirty. This is needed by find-all-properties-references. When we want to find all references to a property, we can dirtify the definition that contains the property, and then the recheck infra will automatically compute the recheck set.

e.g. in
```
type Foo = { bar: string };
//           ^ loc to dirtify
```
, the definition of `Foo` will be dirtified.

This diff computes which definitions will be dirtified. It doesn't attempt to change the signature hash yet, which will be done in a future diff.

On the implementation side, we compute the definitions that need to be dirtified during the marking phase of type sig generation. Everytime we visit a location, we check whether the location needs to be dirtified, if so, we will record the fact to the associated definition of the location. In the end, we will record the indices of the compacted definition array into separate arrays.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D45756710

fbshipit-source-id: 54cff7d234fcbe0c5716448df257e8c8f5dcce73