[IDE] Boilerplate for RenameFileImports action
Summary:
This diff sets up the boilerplate for renaming all file imports when a file is renamed. Since we want this behavior only after confirming the action with the user on the VSCode extension side (in D44775021), I implemented it in a custom action, `flow/renameFileImports`. Behavior will be added in subsequent diffs.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D44775026

fbshipit-source-id: e30285ad49946ca0082bbe17edfc7f56ee3f2699