Add prefix field to file_sig to preserve original require string
Summary:
This diff adds the string prefix into the file_sig record for `Require`. This extra information will be necessary when using the File_sig to find requires for find/replace. When replacing module names, the stored prefix can now be prepended to the new module name to correctly reproduce the original require format.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D44678814

fbshipit-source-id: 14b7fe6e82b5ed76ea137f8cb6b0f2568e38e535