improve predicate function errors on pattern and rest params
Summary:
Predicate functions on a destructured or rest parameter are not supported. But there is
no reason why we can't support predicate functions with destructuring patterns or rest
on parameters other than the one that is used for refinement. For example
```
function foo({a: b}: {a:mixed}, x: mixed, ...rest: Array<mixed>): %checks {
  return typeof x === "number";
}
```
Today, this is not allowed since we wholesale ban patterns/rest with predicate functions.

This diff makes this change by tracking closely the exact elements names that are bound
in the parameters, while also storing the binding information for each one. For example, for
the above we'd store
* `x` -> Root
* `b` -> Select('a', Root)
* `rest` -> Rest

Then after we finally determine the predicate maps encoded in the return expression, we
enforce that the domain of these maps (the refined parameters) are all of kind Root.
Otherwise, we raise an appropriate message.

I'm planning to apply the same kind of check for type guard functions.

This diff introduces a Ast.Pattern.t visitor in pattern_helper.ml. This visitor is similar to the
one found in Name_def, but sharing this code wasn't very practical.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D45805400

fbshipit-source-id: e647715577554b4e31d28bed7a86124b91b543fe